// src/adjustment/domain/enums/adjustment.type.enums.ts
// export enum AdjustmentTypeEnums {
//     ADJUST_DOWN = 0,
//     ADJUST_UP = 1,
//     ADJUST_INIT = 2
// }

export enum AdjustmentTypeEnums {
    DOWN = "DOWN",
    UP = "UP",
    INIT = "INIT"
}

export enum AdjustmentActionEnums {
    INIT = "INIT",
    ISSUE = "ISSUE",
    PROCUREMENT = "PROCUREMENT",
    RECEIPT = "RECEIPT",
    SALE = "SALE"
}